drive-by branch that splits click.go into meaningful files
